refactor(runner): split daemon health classifier
